\name{USlobbying-package}
\alias{USlobbying-package}
\alias{USlobbying}
\docType{package}
\title{
Documents Filed under the US Lobbying Disclosure Act
}
\description{
  Download and clean documents filed under the Lobbying Disclosure Act
  (LDA) of the United States of America.
}
\details{
  \tabular{ll}{
    Package: \tab USlobbying\cr
    Type: \tab Package\cr
    Version: \tab 0.0-0\cr
    Date: \tab 2012-03-30\cr
    License: \tab GPL (>= 2)\cr
    URL: \tab http://www.r-project.org\cr
    Depends: \tab \cr
    Suggests: \tab \cr
    Packaged: \tab \cr
    Built: \tab \cr
  }

  Index:
  \preformatted{
  }

  See the \code{scripts} subdirectory of the \code{FinTS} installation
  directory =  system.file(package='FinTS').

  Corrections to the script files provided and contributions to script
  files for other chapters will be graciously accepted.
}
\author{
Spencer Graves

Maintainer: Spencer Graves <spencer.graves@prodsyse.com>
}
%\references{}
\keyword{ package }
\keyword{ IO }
\seealso{
  \url{
http://www.senate.gov/legislative/Public_Disclosure/LDA_reports.htm
  }
  \url{http://rootstrikers.org}
  \url{http://unitedrepublic.org}
}
%\examples{}
